+++ src/main/things/Birch.java
@@ -24,7 +24,7 @@ public class Birch extends Tree {
 
 	@Override
 	public void drop(Vector<Item> items,ImageContainer container) {
-		items.add(Functions.createItem(5, 1, container, (int) this.angle));
+		items.add(Functions.createItem(5, 1, container, this.angle));
 	}
 
 }
