+++ src/main/things/Spruce.java
@@ -25,7 +25,7 @@ public class Spruce extends Tree {
 
 	@Override
 	public void drop(Vector<Item> items,ImageContainer container) {
-		items.add(Functions.createItem(5, 1, container, (int) this.angle));
+		items.add(Functions.createItem(5, 1, container, this.angle));
 	}
 
 }
